refactor: extract bundle spatial transforms
