+++ bpy_speckle/converter/from_bundle/bundle_to_native.py
@@ -22,13 +22,14 @@ from typing import Any, Dict, List, Optional, Tuple
 import bpy
 from mathutils import Matrix
 from specklepy.bundle import sgeo
-from specklepy.objects.models.units import (
-    get_scale_factor_to_meters,
-    get_units_from_string,
-)
-
 from ..utils import create_material_from_proxy
 from ._baking.result import BakeResult
+from ._baking.transforms import (
+    origin_median,
+    placement_matrix,
+    recenter_origin,
+    scale_for,
+)
 from .bundle_reader import (
     BundleGeometry,
     BundleMaterial,
@@ -52,14 +53,6 @@ _DECODABLE_TYPES = _MESH_TYPES | _CURVE_TYPES | _POINT_TYPES
 # segments used when flattening an analytical arc/circle/ellipse to a polyline.
 # Blender has no native arc primitive, so these are tessellated on the way in.
 _ARC_SEGMENTS = 64
-
-
-def _scale_for(units: Optional[str]) -> float:
-    """Scale factor from a bundle unit string into the Blender scene."""
-    if not units:
-        return 1.0
-    unit_scale = get_scale_factor_to_meters(get_units_from_string(units))
-    return unit_scale / bpy.context.scene.unit_settings.scale_length
 
 
 class _MaterialShim:
@@ -329,7 +322,7 @@ def _add_splines(curve_data: bpy.types.Curve, curve) -> None:
             _add_splines(curve_data, segment)
         return
 
-    scale = _scale_for(getattr(curve, "units", None))
+    scale = scale_for(getattr(curve, "units", None))
 
     if kind == "Curve":
         _add_nurbs_spline(curve_data, curve, scale)
@@ -537,7 +530,7 @@ def _points_object(name: str, points: List[object]) -> Optional[bpy.types.Object
     """
     coords: List[Tuple[float, float, float]] = []
     for point in points:
-        scale = _scale_for(getattr(point, "units", None))
+        scale = scale_for(getattr(point, "units", None))
         if type(point).__name__ == "PointCloud":
             coords.extend((p.x * scale, p.y * scale, p.z * scale) for p in point.points)
         else:
@@ -580,7 +573,7 @@ def _mesh_datablock(
     current_face = 0
 
     for mesh_index, mesh in enumerate(meshes):
-        scale = _scale_for(mesh.units)
+        scale = scale_for(mesh.units)
         vertices = mesh.vertices
         for i in range(0, len(vertices), 3):
             all_vertices.append(
@@ -740,22 +733,6 @@ def _apply_properties(
     result.dropped_properties += dropped
 
 
-def _placement_matrix(transform: List[float], units: Optional[str]) -> Matrix:
-    """Turn a placement's 16 row-major doubles into a Blender matrix.
-
-    The translation is unit-scaled but the rotation/shear block is not — scaling
-    the whole matrix would scale the basis vectors too and shrink the instance.
-    """
-    if len(transform) != 16:
-        return Matrix.Identity(4)
-    matrix = Matrix([transform[0:4], transform[4:8], transform[8:12], transform[12:16]])
-    scale = _scale_for(units)
-    if scale != 1.0:
-        for row in range(3):
-            matrix[row][3] *= scale
-    return matrix
-
-
 # ── orchestration ───────────────────────────────────────────────────────────
 
 
@@ -886,81 +863,6 @@ def _member_objects(
     )
 
 
-def _local_bounds_center(data: object) -> Optional[Tuple[float, float, float]]:
-    """Bounds center of a mesh or curve data-block, in its own coordinates.
-
-    Computed from the data directly — ``Object.bound_box`` is only valid after
-    a depsgraph evaluation, which freshly created objects have not had.
-    """
-    if isinstance(data, bpy.types.Mesh):
-        count = len(data.vertices)
-        if not count:
-            return None
-        flat = [0.0] * (count * 3)
-        data.vertices.foreach_get("co", flat)
-        xs, ys, zs = flat[0::3], flat[1::3], flat[2::3]
-    elif isinstance(data, bpy.types.Curve):
-        xs, ys, zs = [], [], []
-        for spline in data.splines:
-            points = spline.bezier_points if spline.type == "BEZIER" else spline.points
-            for point in points:
-                co = point.co
-                xs.append(co[0])
-                ys.append(co[1])
-                zs.append(co[2])
-        if not xs:
-            return None
-    else:
-        return None
-    return (
-        (min(xs) + max(xs)) / 2.0,
-        (min(ys) + max(ys)) / 2.0,
-        (min(zs) + max(zs)) / 2.0,
-    )
-
-
-def _recenter_origin(obj: bpy.types.Object) -> None:
-    """Move a world-baked object's origin onto its geometry's bounds center.
-
-    Direct-baked data carries world coordinates under an identity matrix, so
-    every object's origin sits at (0, 0, 0). That is invisible until the object
-    joins a parent relationship: Blender draws relationship lines
-    origin-to-origin, so a placed child linked to an origin-bound parent draws
-    a line across the whole scene. Only parenting participants are recentred;
-    everything else keeps the identity transform the dialect promises.
-
-    World geometry is preserved exactly — the data shifts by ``-center`` and
-    the matrix gains ``+center``. Recentring twice is a no-op (the second
-    center is ~zero), and shared data is left alone: shifting it would move
-    every other user.
-    """
-    data = obj.data
-    if data is None or data.users > 1:
-        return
-    center = _local_bounds_center(data)
-    if center is None or max(abs(c) for c in center) < 1e-9:
-        return
-    data.transform(Matrix.Translation((-center[0], -center[1], -center[2])))
-    obj.matrix_world = obj.matrix_world @ Matrix.Translation(center)
-
-
-def _origin_median(origins: List[object]) -> Tuple[float, float, float]:
-    """Component-wise median — resists the one far-flung subelement."""
-
-    def median(values: List[float]) -> float:
-        values = sorted(values)
-        mid = len(values) // 2
-        if len(values) % 2:
-            return values[mid]
-        return (values[mid - 1] + values[mid]) / 2.0
-
-    return (
-        median([o[0] for o in origins]),
-        median([o[1] for o in origins]),
-        median([o[2] for o in origins]),
-    )
-
-
 def _objects_from_geometries(
     name: str,
     data_name: str,
@@ -1021,7 +923,7 @@ def _objects_from_geometries(
     # extras' world-space data stays put.
     if len(built) > 1:
         for part in built:
-            _recenter_origin(part)
+            recenter_origin(part)
         primary_inverse = built[0].matrix_world.inverted(Matrix.Identity(4))
         for extra in built[1:]:
             extra.parent = built[0]
@@ -1104,7 +1006,7 @@ def _build_definitions(
             empty = bpy.data.objects.new(f"{nested.name}.{ordinal}", None)
             empty.instance_type = "COLLECTION"
             empty.instance_collection = nested
-            empty.matrix_world = _placement_matrix(instance.transform, instance.units)
+            empty.matrix_world = placement_matrix(instance.transform, instance.units)
             definition_collections[node_id].objects.link(empty)
 
     return definition_collections
@@ -1144,7 +1046,7 @@ def _bake_placement(
         definition = definition_collections.get(instance.def_ref)
         if definition is None:
             continue
-        matrix = _placement_matrix(instance.transform, instance.units)
+        matrix = placement_matrix(instance.transform, instance.units)
         if instance_loading_mode == "LINKED_DUPLICATES":
             built.extend(_duplicate_definition(name, definition, matrix, frozenset()))
         else:
@@ -1253,17 +1155,17 @@ def _parent_subelements(bundle: ReceivedBundle, result: BakeResult) -> None:
             if child in anchored or (
                 child_obj and (child_obj.is_placement or child_obj.geometry_ks)
             ):
-                _recenter_origin(child)
+                recenter_origin(child)
                 placed.append((child, child.matrix_world.copy()))
             else:
                 followers.append(child)
 
         # the parent's origin must be final before any child is linked: a
         # child's world restore resolves against the parent's stored matrix
-        _recenter_origin(parent)
+        recenter_origin(parent)
         if not obj.is_placement and not obj.geometry_ks and placed:
             parent.matrix_world = Matrix.Translation(
-                _origin_median([world.to_translation() for _, world in placed])
+                origin_median([world.to_translation() for _, world in placed])
             )
             anchored.add(parent)
